# internal/entities.go
package internal

type User struct {
	Id       int    `json:"id" db:"id"`
	Username string `json:"username" db:"username"`
	Email    string `json:"email" db:"email"`
}

type errorResponse struct {
	Message string `json:"message"`
}